requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot==20.7
python-dotenv==1.0.0
